chore: clear the 14 golangci-lint findings CI caught
Local gates never ran golangci-lint, so the wave shipped errcheck,
errorlint, gosec, govet, and staticcheck findings: unchecked test
WriteStrings, a non-wrapping %v beside a %w, a 0o644 export, an
unbounded uint64 shift in poll backoff (shift is already clamped —
the conversion now says so), math/rand jitter (annotated, it is not
cryptographic), three int/uint32 uid conversions now compared as
int64 with the reflect Uint bounded, reflect.Ptr -> reflect.Pointer,
a mergeable conditional assignment in the gate, and two untagged
switches in doctor. golangci-lint run: 0 issues.

+++ internal/delivery/delivery.go
@@ -642,7 +642,7 @@ func NextCheck(now time.Time, attempt int, statusPollMinutes int) time.Time {
 		// multiplication for a pathologically large attempt count.
 		shift = 10
 	}
-	backoff := time.Duration(base) * time.Minute * time.Duration(uint64(1)<<uint(shift))
+	backoff := time.Duration(base) * time.Minute * time.Duration(int64(1)<<shift)
 	if backoff <= 0 || backoff > maxPollInterval {
 		backoff = maxPollInterval
 	}
